Split all classes, structs and interfaces inside of Tank.cs to separate files.
